+++ tokio-util/Cargo.toml
@@ -35,8 +35,7 @@ rt = ["tokio/rt"]
 __docs_rs = ["futures-util"]
 
 [dependencies]
-#tokio = { version = "1.0.0", features = ["sync"] }
-tokio = { path = "../tokio", features = ["full"] }
+tokio = { version = "1.0.0", features = ["sync"] }
 
 bytes = "1.0.0"
 futures-core = "0.3.0"
@@ -48,6 +47,7 @@ pin-project-lite = "0.2.0"
 slab = { version = "0.4.1", optional = true } # Backs `DelayQueue`
 
 [dev-dependencies]
+tokio = { version = "1.0.0", features = ["sync"] }
 tokio-test = { version = "0.4.0" }
 tokio-stream = { version = "0.1" }
