+++ tokio-util/src/sync/mod.rs
@@ -13,6 +13,3 @@ pub use poll_semaphore::PollSemaphore;
 
 mod reusable_box;
 pub use reusable_box::ReusableBoxFuture;
-
-mod condvar;
-pub use condvar::Condvar;
